first steps to get contacts from db8.
Not yet working.. Not sure, why... fetch in db8Source is never called.
hm.

I implemented a db8Source which should be able to
fetch/commit/find/destroy models and collections from/to/in db8.

I connected that to a Model and a Collection...

We currently don't have the contact linker running/working, so
getting contacts from com.palm.contact:1 instead of person.

In future that should change to com.palm.person:1 for reading,
but still into the contacts kinds for writing.. I actually never
really understood how webOS is handling this internally. Need to
investigate that a bit further.